+++ plugin/gnome/NetworkManagerGnomeUtils.cpp
@@ -250,16 +250,11 @@ namespace WPEFramework
             return true;
         }
 
-        bool nmUtils::configureNetworkManagerDaemonLoglevel()
+        bool nmUtils::setNetworkManagerlogLevelToTrace()
         {
             /* set networkmanager daemon log level based on current plugin log level */
             const char* command = "nmcli general logging level TRACE domains ALL";
-            NetworkManagerLogger::LogLevel level;
-            NetworkManagerLogger::GetLevel(level);
 
-            if(NetworkManagerLogger::DEBUG_LEVEL != level)
-                return false;
- 
             // Execute the command using popen
             FILE *pipe = popen(command, "r");
 
@@ -267,7 +262,7 @@ namespace WPEFramework
                 NMLOG_ERROR("popen failed %s ", command);
                 return false;
             }
-    
+
             // Close the pipe and retrieve the command's exit status
             int status = pclose(pipe);
             if (status == -1) {
@@ -278,12 +273,11 @@ namespace WPEFramework
             // Extract the exit status from the status code
             int exitCode = WEXITSTATUS(status);
             if (exitCode == 0)
-                NMLOG_INFO("NetworkManager daemon log level changed ! logLevel: %d", level);
+                NMLOG_INFO("NetworkManager daemon log level changed to trace");
             else
                 NMLOG_INFO(" '%s' failed with exit code %d.", command, exitCode);
 
             return exitCode == 0;
         }
-
     }   // Plugin
 }   // WPEFramework